requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
pyrebase4==4.7.1
